feat(resolve): allow overriding enhanced-resolve's options
