+++ docs/shared-settings.md
@@ -60,28 +60,24 @@ If a path is relative, it will be resolved from CWD.
 { "resolvePaths": [] }
 ```
 
-## resolveModules
+## resolverConfig
 
-Adds additional directories that should be searched when resolving modules.
+Override the options generated by this plugin for resolving require and import statements. These
+options are passed down to `enhanced-resolve`'s factory method and are documented [here on their
+GitHub repository](https://github.com/webpack/enhanced-resolve?tab=readme-ov-file#resolver-options).
 
-Absolute and relative paths can both be used, but be aware that they will behave a bit differently.
+The options you define here are assigned over the default options generated by the plugin.
 
-A relative path will be scanned similarly to how Node scans for node_modules, by looking through the current directory as well as its ancestors (i.e. ./node_modules, ../node_modules, and on).
-
-With an absolute path, it will only search in the given directory.
-
-If a path is relative, it will be resolved from CWD.
-
-### Example resolveModules
+### Example resolverConfig
 
 ```json
-{ "resolveModules": ["/path/to/somewhere", "../relative/path", "node_modules"] }
+{ "resolverConfig": { "modules": ["node_modules", "bower_components"] } }
 ```
 
-### Default resolveModules
+### Default resolverConfig
 
 ```json
-{ "resolveModules": ["node_modules"] }
+{ "resolverConfig": {} }
 ```
 
 ## convertPath
