+++ lib/rules/no-missing-import.js
@@ -7,7 +7,7 @@
 const { checkExistence, messages } = require("../util/check-existence")
 const getAllowModules = require("../util/get-allow-modules")
 const getResolvePaths = require("../util/get-resolve-paths")
-const getResolveModules = require("../util/get-resolve-modules")
+const getResolverConfig = require("../util/get-resolver-config")
 const getTryExtensions = require("../util/get-try-extensions")
 const getTSConfig = require("../util/get-tsconfig")
 const getTypescriptExtensionMap = require("../util/get-typescript-extension-map")
@@ -30,7 +30,7 @@ module.exports = {
                 properties: {
                     allowModules: getAllowModules.schema,
                     resolvePaths: getResolvePaths.schema,
-                    resolveModules: getResolveModules.schema,
+                    resolverConfig: getResolverConfig.schema,
                     tryExtensions: getTryExtensions.schema,
                     ignoreTypeImport: { type: "boolean", default: false },
                     tsconfigPath: getTSConfig.schema,
