+++ lib/rules/no-missing-require.js
@@ -7,7 +7,7 @@
 const { checkExistence, messages } = require("../util/check-existence")
 const getAllowModules = require("../util/get-allow-modules")
 const getResolvePaths = require("../util/get-resolve-paths")
-const getResolveModules = require("../util/get-resolve-modules")
+const getResolverConfig = require("../util/get-resolver-config")
 const getTSConfig = require("../util/get-tsconfig")
 const getTryExtensions = require("../util/get-try-extensions")
 const getTypescriptExtensionMap = require("../util/get-typescript-extension-map")
@@ -31,7 +31,7 @@ module.exports = {
                     allowModules: getAllowModules.schema,
                     tryExtensions: getTryExtensions.schema,
                     resolvePaths: getResolvePaths.schema,
-                    resolveModules: getResolveModules.schema,
+                    resolverConfig: getResolverConfig.schema,
                     typescriptExtensionMap: getTypescriptExtensionMap.schema,
                     tsconfigPath: getTSConfig.schema,
                 },
