+++ lib/rules/no-unpublished-import.js
@@ -8,7 +8,7 @@ const { checkPublish, messages } = require("../util/check-publish")
 const getAllowModules = require("../util/get-allow-modules")
 const getConvertPath = require("../util/get-convert-path")
 const getResolvePaths = require("../util/get-resolve-paths")
-const getResolveModules = require("../util/get-resolve-modules")
+const getResolverConfig = require("../util/get-resolver-config")
 const visitImport = require("../util/visit-import")
 
 /** @type {import('eslint').Rule.RuleModule} */
@@ -29,7 +29,7 @@ module.exports = {
                     allowModules: getAllowModules.schema,
                     convertPath: getConvertPath.schema,
                     resolvePaths: getResolvePaths.schema,
-                    resolveModules: getResolveModules.schema,
+                    resolverConfig: getResolverConfig.schema,
                     ignoreTypeImport: { type: "boolean", default: false },
                     ignorePrivate: { type: "boolean", default: true },
                 },
