+++ lib/rules/no-unpublished-require.js
@@ -8,7 +8,7 @@ const { checkPublish, messages } = require("../util/check-publish")
 const getAllowModules = require("../util/get-allow-modules")
 const getConvertPath = require("../util/get-convert-path")
 const getResolvePaths = require("../util/get-resolve-paths")
-const getResolveModules = require("../util/get-resolve-modules")
+const getResolverConfig = require("../util/get-resolver-config")
 const getTryExtensions = require("../util/get-try-extensions")
 const visitRequire = require("../util/visit-require")
 
@@ -30,7 +30,7 @@ module.exports = {
                     allowModules: getAllowModules.schema,
                     convertPath: getConvertPath.schema,
                     resolvePaths: getResolvePaths.schema,
-                    resolveModules: getResolveModules.schema,
+                    resolverConfig: getResolverConfig.schema,
                     tryExtensions: getTryExtensions.schema,
                     ignorePrivate: { type: "boolean", default: true },
                 },
