+++ lib/util/import-target.js
@@ -71,7 +71,7 @@ function getTSConfigAliases(context) {
  * @typedef Options
  * @property {string[]} [extensions]
  * @property {string[]} [paths]
- * @property {string[]} [modules]
+ * @property {Partial<import('enhanced-resolve').ResolveOptions>} [resolverConfig]
  * @property {string} basedir
  */
 /** @typedef { 'unknown' | 'relative' | 'absolute' | 'node' | 'npm' | 'http' } ModuleType */
@@ -310,15 +310,16 @@ module.exports = class ImportTarget {
             this.resolverConfig.extensions = this.options.extensions
         }
 
-        if (this.options.modules) {
-            this.resolverConfig.modules = this.options.modules
-        }
-
         if (isTypescript(this.context)) {
             this.resolverConfig.alias = getTSConfigAliases(this.context)
             this.resolverConfig.extensionAlias = getTypescriptExtensionMap(
                 this.context
             ).backward
+        }
+
+        this.resolverConfig = {
+            ...this.resolverConfig,
+            ...this.options.resolverConfig,
         }
 
         const requireResolve = resolver.create.sync(this.resolverConfig)
