+++ lib/util/visit-import.js
@@ -7,7 +7,7 @@
 const path = require("path")
 const { isBuiltin } = require("node:module")
 const getResolvePaths = require("./get-resolve-paths")
-const getResolveModules = require("./get-resolve-modules")
+const getResolverConfig = require("./get-resolver-config")
 const getTryExtensions = require("./get-try-extensions")
 const ImportTarget = require("./import-target")
 const stripImportPathParams = require("./strip-import-path-params")
@@ -42,9 +42,9 @@ module.exports = function visitImport(
         path.resolve(context.filename ?? context.getFilename())
     )
     const paths = getResolvePaths(context, optionIndex)
-    const modules = getResolveModules(context, optionIndex)
+    const resolverConfig = getResolverConfig(context, optionIndex)
     const extensions = getTryExtensions(context, optionIndex)
-    const options = { basedir, paths, modules, extensions }
+    const options = { basedir, paths, extensions, resolverConfig }
 
     /**
      * @param {(
