+++ lib/util/visit-require.js
@@ -12,7 +12,7 @@ const {
 } = require("@eslint-community/eslint-utils")
 const { isBuiltin } = require("node:module")
 const getResolvePaths = require("./get-resolve-paths")
-const getResolveModules = require("./get-resolve-modules")
+const getResolverConfig = require("./get-resolver-config")
 const getTryExtensions = require("./get-try-extensions")
 const ImportTarget = require("./import-target")
 const stripImportPathParams = require("./strip-import-path-params")
@@ -43,9 +43,9 @@ module.exports = function visitRequire(
         path.resolve(context.filename ?? context.getFilename())
     )
     const paths = getResolvePaths(context)
-    const modules = getResolveModules(context)
+    const resolverConfig = getResolverConfig(context)
     const extensions = getTryExtensions(context)
-    const options = { basedir, paths, modules, extensions }
+    const options = { basedir, paths, extensions, resolverConfig }
 
     return {
         "Program:exit"(node) {
